move kueue utils
