+++ frontend/src/concepts/hardwareProfiles/kueueConstants.ts
