+++ frontend/src/concepts/hardwareProfiles/kueueUtils.ts
